requirements.txt
paho-mqtt~=1.6.1
Pillow~=9.0.0
python-dotenv
requests~=2.27.1
urllib3~=1.26.8
